Extract the product category list view and view model to its own class.
