+++ WooCommerce/Classes/ViewRelated/Products/Categories/ProductCategoryCellViewModel.swift
