+++ WooCommerce/Classes/ViewRelated/Products/Categories/ProductCategoryListViewController.swift
@@ -6,24 +6,17 @@ import WordPressUI
 ///
 final class ProductCategoryListViewController: UIViewController {
 
-    @IBOutlet private weak var addButton: UIButton!
-    @IBOutlet private weak var addButtonSeparator: UIView!
-    @IBOutlet private var tableView: UITableView!
+    @IBOutlet var tableView: UITableView!
     private let ghostTableView = UITableView()
 
-    private let viewModel: ProductCategoryListViewModel
+    let viewModel: ProductCategoryListViewModel
 
     private let siteID: Int64
 
-    // Completion callback
-    //
-    typealias Completion = (_ categories: [ProductCategory]) -> Void
-    private let onCompletion: Completion
+    init(siteID: Int64) {
+        self.viewModel = ProductCategoryListViewModel(siteID: siteID)
+        self.siteID = siteID
 
-    init(product: Product, completion: @escaping Completion) {
-        self.viewModel = ProductCategoryListViewModel(product: product)
-        self.siteID = product.siteID
-        onCompletion = completion
         super.init(nibName: type(of: self).nibName, bundle: nil)
     }
 
@@ -33,12 +26,10 @@ final class ProductCategoryListViewController: UIViewController {
 
     override func viewDidLoad() {
         super.viewDidLoad()
-        configureAddButton()
-        configureAddButtonSeparator()
+
         registerTableViewCells()
         configureTableView()
         configureGhostTableView()
-        configureNavigationBar()
         configureViewModel()
         handleSwipeBackGesture()
     }
@@ -50,16 +41,6 @@ private extension ProductCategoryListViewController {
     func registerTableViewCells() {
         tableView.registerNib(for: ProductCategoryTableViewCell.self)
         ghostTableView.registerNib(for: ProductCategoryTableViewCell.self)
-    }
-
-    func configureAddButton() {
-        addButton.setTitle(NSLocalizedString("Add Category", comment: "Action to add category on the Product Categories screen"), for: .normal)
-        addButton.addTarget(self, action: #selector(addButtonTapped), for: .touchUpInside)
-        addButton.applySecondaryButtonStyle()
-    }
-
-    func configureAddButtonSeparator() {
-        addButtonSeparator.backgroundColor = .systemColor(.separator)
     }
 
     func configureTableView() {
@@ -77,25 +58,6 @@ private extension ProductCategoryListViewController {
         ghostTableView.pinSubviewToAllEdges(view)
         ghostTableView.backgroundColor = .listBackground
         ghostTableView.removeLastCellSeparator()
-    }
-
-    func configureNavigationBar() {
-        configureTitle()
-        configureRightButton()
-    }
-
-    func configureTitle() {
-        title = NSLocalizedString("Categories", comment: "Edit product categories screen - Screen title")
-    }
-
-    func configureRightButton() {
-        let applyButtonTitle = NSLocalizedString("Done",
-                                               comment: "Edit product categories screen - button title to apply categories selection")
-        let rightBarButton = UIBarButtonItem(title: applyButtonTitle,
-                                             style: .done,
-                                             target: self,
-                                             action: #selector(doneButtonTapped))
-        navigationItem.setRightBarButton(rightBarButton, animated: false)
     }
 }
 
@@ -117,47 +79,6 @@ private extension ProductCategoryListViewController {
                 self?.removeGhostTableView()
             }
         }
-    }
-}
-
-// MARK: - Navigation actions handling
-//
-extension ProductCategoryListViewController {
-
-    override func shouldPopOnBackButton() -> Bool {
-        if viewModel.hasUnsavedChanges() {
-            presentBackNavigationActionSheet()
-            return false
-        }
-        return true
-    }
-
-    override func shouldPopOnSwipeBack() -> Bool {
-        return shouldPopOnBackButton()
-    }
-
-    @objc private func doneButtonTapped() {
-        ServiceLocator.analytics.track(.productCategorySettingsDoneButtonTapped)
-        onCompletion(viewModel.selectedCategories)
-    }
-
-    @objc private func addButtonTapped() {
-        ServiceLocator.analytics.track(.productCategorySettingsAddButtonTapped)
-        let addCategoryViewController = AddProductCategoryViewController(siteID: siteID) { [weak self] (newCategory) in
-            defer {
-                self?.dismiss(animated: true, completion: nil)
-            }
-            self?.viewModel.addAndSelectNewCategory(category: newCategory)
-            self?.tableView.reloadData()
-        }
-        let navController = WooNavigationController(rootViewController: addCategoryViewController)
-        present(navController, animated: true, completion: nil)
-    }
-
-    private func presentBackNavigationActionSheet() {
-        UIAlertController.presentDiscardChangesActionSheet(viewController: self, onDiscard: { [weak self] in
-            self?.navigationController?.popViewController(animated: true)
-        })
     }
 }
