+++ WooCommerce/Classes/ViewRelated/Products/Categories/ProductCategoryViewModelBuilder.swift
