+++ WooCommerce/WooCommerceTests/ViewRelated/Products/Categories/ProductCategoryViewModelBuilderTests.swift
